REFACTOR: token generation and import statements
