+++ server/models/token.js
@@ -1,9 +1,9 @@
 const mongoose = require('mongoose')
+const tokenHelper = require('../utils/token')
 
 const tokenSchema = new mongoose.Schema({
   token: {
-    type: String,
-    required: true,
+    type: String
   },
   user: {
     type: mongoose.Schema.Types.ObjectId,
@@ -23,6 +23,14 @@ const tokenSchema = new mongoose.Schema({
     enum: ['verify', 'reset'],
     required: true,
   },
+})
+
+// if token is not given in the request during create, generate a new one 
+tokenSchema.pre('save', function (next) {
+  if (!this.token) {
+    this.token = tokenHelper.generateOneTimeToken()
+  }
+  next()
 })
 
 module.exports = mongoose.model('Token', tokenSchema)
